Implemented the DataSender as a virtual class.
The MQTT connection is done by MQTTDataSender which now inherits from DataSender.
